Use normal chan messages for public commands instead of private notices
For public bot commands (e.g., “botchat”) issued in a channel, send
feedback messages to the channel instead of directly to the user.
This is because SAX users in #cemetech can't read private messages, and
IRC users sometimes overlook them (or have PMs blocked).
(Messages for privately-issued commands outside of channels should of
course still be PMed to the user.)

Use normal messages now instead of notices because the bot is a chatbot
that does most of its talking via normal messages, anyway.  Might as well
be consistent.