Hack around a race condition when generating new personalities
get_corpus.py and train.py get run in parallel, with the former piping
output to the latter.  Unfortunately, they both initialize a
PostgresMarkov object for the same personality, which attempts to use
CREATE TABLE IF NOT EXISTS to generate the same tables at the same
time… and evidently this Postgres command is not atomic.

So for now, just add a parameter to the class to supress the table
creation and use it in get_corpus.py, since train.py will do it and
get_corpus.py doesn't seem to use the tables in question anyway.