Support a relay named BN-Relay and refactor control-char sanitization
The control-character stripping was in the NikkyAI class and module,
and it logically doesn't belong there. It's an IRC-specific thing, so
it should have been in nikkybot.py. So, in order to properly handle
the relay's use of color formatting for the nickname, move the
function to the appropriate module as well.