Yet another attempt to make flood protection work how I'd truly like
This time there's an additional “cooldown” timer to force the flooder
to shut up for a while before it ever responds to them again. Once the
maximum number of messages per interval is exceeded, the cooldown
timer will start, and it resets if the flooder attempts to talk to the
bot again too early. Only once the timer expires with no attempted
messages in the meantime will the flooder stop being ignored.