Work around decoding error crashes with IRC log parsing
We'll assume everything is UTF-8 and replace unacceptable characters
with backslash references. It's not perfect because in reality some of
my older logs are a bloody mess of encodings, some badly corrupted and
disfigured, but trying to fix *that* is way out of the scope of this
project. This way we at least run to completion and are not much worse
off than Python 2 which blindly ignored encodings altogether.